test/pil/requirements.txt
python-can>=4.2.0
pyyaml>=6.0
junit-xml>=1.9
